+++ src/math/matrix_operations.rs
@@ -2,13 +2,20 @@ use super::matrix::Matrix;
 use super::errors::*;
 
 impl Matrix {
-
     pub fn add(first: &Matrix, second: &Matrix) -> MathResult<Self> {
         Self::element_wise(first, second, |x, y| x + y)
     }
 
     pub fn sub(first: &Matrix, second: &Matrix) -> MathResult<Self> {
         Self::element_wise(first, second, |x, y| x - y)
+    }
+
+    pub fn mul(first: &Matrix, second: &Matrix) -> MathResult<Self> {
+        Self::element_wise(first, second, |x, y| x * y)
+    }
+
+    pub fn div(first: &Matrix, second: &Matrix) -> MathResult<Self> {
+        Self::element_wise(first, second, |x, y| x / y)
     }
 
     pub fn product(first: &Matrix, second: &Matrix) -> MathResult<Self> {
@@ -31,13 +38,28 @@ impl Matrix {
         }        
     }
 
-    pub fn transposed(&self) -> Matrix {
+    pub fn transpose(&self) -> Matrix {
         let (rows, cols) = (self.rows(), self.cols());
         Matrix::new(cols, rows, |i, j| {
             self.get_unchecked(j, i)
         })
     }
 
+    pub fn plus_assign(&mut self, other: &Matrix) -> MathResult<&mut Self> {
+        self.element_wise_other(other, |a, b| a + b)        
+    }
+
+    pub fn minus_assign(&mut self, other: &Matrix) -> MathResult<&mut Self> {
+        self.element_wise_other(other, |a, b| a - b)
+    }
+
+    pub fn multiplicate_assign(&mut self, scalar: f64) -> &mut Self {
+        self.element_wise_mut(|x| x * scalar)
+    }
+
+    pub fn divide_assign(&mut self, scalar: f64) -> &mut Self {
+        self.element_wise_mut(|x| x / scalar)
+    }
 }
 
 #[cfg(test)]
@@ -45,7 +67,7 @@ mod tests {
     use super::*;
 
     #[test]
-    fn matrix_operation_plus() -> MathResult<()> {
+    fn matrix_operation_add() -> MathResult<()> {
         let m1 = Matrix::from_vector(&vec![
             vec![1.0, 2.0, 3.0],
             vec![4.0, 5.0, 6.0],
@@ -70,7 +92,7 @@ mod tests {
     }
 
     #[test]
-    fn matrix_operation_minus() -> MathResult<()> {
+    fn matrix_operation_sub() -> MathResult<()> {
         let m1 = Matrix::from_vector(&vec![
             vec![1.0, 2.0, 3.0],
             vec![4.0, 5.0, 6.0],
@@ -170,7 +192,7 @@ mod tests {
     }
 
     #[test]
-    fn matrix_operation_transposed() -> MathResult<()> {
+    fn matrix_operation_transpose() -> MathResult<()> {
         let a = Matrix::from_vector(&vec![
             vec![1.0, 2.0, 3.0],
             vec![4.0, 5.0, 6.0]
@@ -182,7 +204,93 @@ mod tests {
             vec![3.0, 6.0]
 
         ])?;
-        assert!(a.transposed() == expected, "Matrix transpose implemented incorrectly");
+        assert!(a.transpose() == expected, "Matrix transpose implemented incorrectly");
+        Ok(())
+    }
+
+    #[test]
+    fn matrix_operation_add_assign() -> MathResult<()> {
+        let mut m1 = Matrix::from_vector(&vec![
+            vec![1.0, 2.0, 3.0],
+            vec![4.0, 5.0, 6.0],
+            vec![7.0, 8.0, 9.0]
+        ])?;
+
+        let m2 = Matrix::from_vector(&vec![
+            vec![9.0, 8.0, 7.0],
+            vec![6.0, 5.0, 4.0],
+            vec![3.0, 2.0, 1.0]
+        ])?;
+
+        let expected = Matrix::from_vector(&vec![
+            vec![10.0, 10.0, 10.0],
+            vec![10.0, 10.0, 10.0],
+            vec![10.0, 10.0, 10.0]
+        ])?;
+
+        m1.plus_assign(&m2)?;
+        assert!(m1 == expected, "Matrix add implemeted incorrectly");
+        Ok(())
+    }
+
+    #[test]
+    fn matrix_operation_sub_assign() -> MathResult<()> {
+        let mut m1 = Matrix::from_vector(&vec![
+            vec![1.0, 2.0, 3.0],
+            vec![4.0, 5.0, 6.0],
+            vec![7.0, 8.0, 9.0]
+        ])?;
+
+        let m2 = Matrix::from_vector(&vec![
+            vec![9.0, 8.0, 7.0],
+            vec![6.0, 5.0, 4.0],
+            vec![3.0, 2.0, 1.0]
+        ])?;
+
+        let expected = Matrix::from_vector(&vec![
+            vec![-8.0, -6.0, -4.0],
+            vec![-2.0,  0.0,  2.0],
+            vec![ 4.0,  6.0,  8.0]
+        ])?;
+
+        m1.minus_assign(&m2)?;
+        assert!(m1 == expected, "Matrix sub implemeted incorrectly");
+        Ok(())
+    }
+
+    #[test]
+    fn matrix_operation_scalar_multiplication() -> MathResult<()> {
+        let mut m = Matrix::from_vector(&vec![
+            vec![1.0, 2.0, 3.0],
+            vec![4.0, 5.0, 6.0],
+            vec![7.0, 8.0, 9.0]
+        ])?;
+
+        let expected = Matrix::from_vector(&vec![
+            vec![ 2.0,  4.0,  6.0],
+            vec![ 8.0, 10.0, 12.0],
+            vec![14.0, 16.0, 18.0]
+        ])?;
+
+        assert!(m.multiplicate_assign(2.0) == &expected, "Matrix scalar multiplication implemented incorrectly");
+        Ok(())
+    }
+
+    #[test]
+    fn matrix_operation_scalar_divide() -> MathResult<()> {
+        let expected = Matrix::from_vector(&vec![
+            vec![1.0, 2.0, 3.0],
+            vec![4.0, 5.0, 6.0],
+            vec![7.0, 8.0, 9.0]
+        ])?;
+
+        let mut m = Matrix::from_vector(&vec![
+            vec![ 2.0,  4.0,  6.0],
+            vec![ 8.0, 10.0, 12.0],
+            vec![14.0, 16.0, 18.0]
+        ])?;
+
+        assert!(m.divide_assign(2.0) == &expected, "Matrix divide by scalar implemented incorrectly");
         Ok(())
     }
