+++ src/math/mod.rs
@@ -6,5 +6,4 @@ pub mod matrix;
 pub mod matrix_debug;
 pub mod matrix_eq;
 pub mod matrix_utils;
-pub mod matrix_mutations;
 pub mod matrix_operations;
